Update carousel component
