+++ frontend/src/common/components/Carousel/types.ts
@@ -7,15 +7,8 @@ export type ChildrenProps = {
   link?: string;
 };
 
-export type CarouselTitle = {
-  heading?: string;
-  description?: string;
-  link?: string;
-};
-
 export type CarouselProps = {
   item: ChildrenProps[];
-  title?: CarouselTitle;
   variant: 'Default' | 'Slider' | 'PopUp';
   fullText: boolean;
   arrows: boolean;
